Added logo, banner and also buttons
Made the notes page have the update and delete notes functions. to
delete hold down on an item and to update just click on it. I guess it
isnt quite consistent with the assignmentss page, but I wanted to added
in the long click because its fun :P
But, I didnt get to the database stuff, because I ran into a big problem
when I refatored the file to change the name.